Chore - Move content from __init__.py to utils.py
